Add Room Attendant page and About Us page done!
