+++ roomranger/src/main/java/org/launchcode/roomranger/controllers/AddAttendantController.java
@@ -1,12 +1,10 @@
 package org.launchcode.roomranger.controllers;
 
-import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import org.launchcode.roomranger.data.ManagerRepository;
 import org.launchcode.roomranger.data.RoomAttendantRepository;
 
 import org.springframework.stereotype.Controller;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.launchcode.roomranger.data.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,14 +12,10 @@ import org.springframework.ui.Model;
 import org.springframework.validation.Errors;
 import org.launchcode.roomranger.models.RoomAttendant;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
-import org.launchcode.roomranger.models.RoomAttendantProfile;
-
-import java.util.List;
-import java.util.Optional;
 
 @Controller
 @RequestMapping(value = "roomAttendant")
-public class RoomAttendantController {
+public class AddAttendantController {
 
     @Autowired
     private RoomAttendantRepository roomAttendantRepository;
@@ -105,26 +99,7 @@ public class RoomAttendantController {
         return "redirect:/roomAttendant";
     }
 
-    @GetMapping("profile")
-    public String getProfile(Model model) {
-        model.addAttribute("title", "Add Room Attendant");
-        model.addAttribute(new RoomAttendant());
-        return "roomAttendant/profile";
-    }
 
-    @PostMapping("/profile/{id}")
-    public String viewProfile(@PathVariable("id") int id, Model model)
-    {
-        RoomAttendant roomAttendant = roomAttendantRepository.findById(id);
-        //List<RoomAttendant> courses = roomAttendant.getRoomAttendant();
-//        if(roomAttendant.isEmpty()) {
-//            return "redirect:/students/" + id + "/addCourses";
-//        }
-       // model.addAttribute("remove_id", id);
-        //roomAttendantRepository.save(roomAttendant);
-        model.addAttribute("roomAttendant", roomAttendant);
-        return "profile";
-    }
 
 }
