+++ roomranger/src/main/resources/templates/roomAttendant/update.html
@@ -12,7 +12,7 @@
 
     <div class="form-group">
       <label>First Name:
-        <input th:text="${roomAttendant.firstName}" class="form-control">
+        <input th:field="${roomAttendant.firstName}" class="form-control">
       </label>
       <p class="error" th:errors="${roomAttendant.firstName}"></p>
     </div>
@@ -62,10 +62,10 @@
     <input  th:field="${roomAttendant.password}" type="password" class="form-control" id="password">
   </div>
 
-              <div>
-                  <label>Upload Picture</label>
-                  <input type="image" src="templates/images" alt="Submit" width="48" height="48">
-              </div>
+  <div>
+    <label>Upload Picture</label>
+    <input type="image" src="templates/images" alt="Submit" width="48" height="48">
+  </div>
   <div class="form-group">
     <label>Notes
       <input th:field="${roomAttendant.notes}" class="form-control">
